requirements.txt
--no-cache-dir
transformers[torch]==4.37.2
torch==2.2.0
accelerate==0.27.2
discord-webhook==1.3.0
python-dotenv==1.0.1
rich==13.7.0 